Remove str(n)cpy from codebase
This function is dangerous: if the resulting string is longer than the
buffer, no null-termination is added to it. Also, it always fills the
remaining buffer length with zeroes.
snprintf is fast and can always be used as replacement (and sometimes
there are better substitutes).

+++ components/agents/posix/application_agent.c
@@ -544,13 +544,29 @@ static int create_unix_domain_socket(const char *path)
 	struct sockaddr_un addr = {
 		.sun_family = AF_UNIX,
 	};
-	strncpy(addr.sun_path, path, sizeof(addr.sun_path) - 1);
+	const int rv1 = snprintf(
+		addr.sun_path,
+		sizeof(addr.sun_path),
+		"%s",
+		path
+	);
+
+	if (rv1 <= 0 || (unsigned int)rv1 > sizeof(addr.sun_path)) {
+		LOGF(
+			"AppAgent: Invalid socket path, len = %d, maxlen = %zu",
+			rv1,
+			sizeof(addr.sun_path)
+		);
+		return -1;
+	}
 
 	unlink(path);
-	int ret = bind(sock,
-		       (const struct sockaddr *) &addr,
-		       sizeof(struct sockaddr_un));
-	if (ret == -1) {
+	int rv2 = bind(
+		sock,
+		(const struct sockaddr *)&addr,
+		sizeof(struct sockaddr_un)
+	);
+	if (rv2 == -1) {
 		LOGERROR("AppAgent", "bind(unix_domain_socket)", errno);
 		return -1;
 	}
